fix: rename lib folder to helpers
